Fixed an error in finds, DBFileInputStream should be used instead of directly using the file

+++ db/bson/tests/testperformance.cpp
@@ -24,7 +24,7 @@
 void testToChar() {
 	BSONObj obj;
 	obj.add("int", 10);
-	obj.add("long", 10L);
+	obj.add("long", (__int64)10L);
 	obj.add("string", std::string("Test"));
 	obj.add("double", 3.14159);
